170d02e6b509b9ae1948374dfeaeb06ad67b94ab

+++ bonus/src/main/java/com/example/bonus/entity/BonusEntity.java
@@ -28,7 +28,7 @@ public class BonusEntity {
 
     private long orderId;
 
-    private int bonusSum;
+    private long bonusSum;
 
     private Instant created;
